third_party/WebKit/LayoutTests/dom/document_type/next-element-sibling.html

<!DOCTYPE html>
<script src="../../resources/js-test.js"></script>
<script>

description('This tests the NonDocumentTypeChildNode interface nextElementSibling property on a DocumentType node.');

var doctype = document.doctype;
shouldBeUndefined('doctype.nextElementSibling');

</script>
