third_party/WebKit/LayoutTests/dom/document_type/next-element-sibling-expected.txt

This tests the NonDocumentTypeChildNode interface nextElementSibling property on a DocumentType node.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS doctype.nextElementSibling is undefined.
PASS successfullyParsed is true

TEST COMPLETE
